Allow giving custom user data to the session created callback function.
This is required because kobby (and libinfinity, apparently) was built
in a way that you join a session, and then you create a new document,
and then a new text buffer on top of that document. Since this is a
KTE plugin now, kate (or whatever application), not us, creates the
document, and we must attach the text buffer to that document
(otherwise we won't see changes). Thus, as creating a session requires
the text buffer, we must know what text buffer to use when
creating a session.

TODO: must check whether this is safe with multiple open requests!!